Refactor Layout.js to Functional Component
