+++ src/frontend/src/components/Layout/index.js
@@ -1,0 +1,3 @@
+import Layout from './Layout.jsx';
+
+export default Layout;
